Update icons used on my account microapp
