--- src/components/ic_description.jsx
@@ -1,19 +1,0 @@
-import { Component } from 'preact'
-import CSS from './styles.css'
-
-export default class Icon extends Component {
-	render() {
-		let color = this.props.color ? CSS[this.props.color] : CSS.active
-		return (
-			<i className={`${CSS.icon} ${color} ${this.props.className || ''}`}>
-				<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-					<path
-						fill="#303334"
-						fillRule="nonzero"
-						d="M14 2l6 6v12c0 1.1-.9 2-2 2H5.99C4.89 22 4 21.1 4 20V4c0-1.1.9-2 2-2h8zm4 18V9h-5V4H6v16h12zM8 16h8v2H8v-2zm0-4h8v2H8v-2z"
-					/>
-				</svg>
-			</i>
-		)
-	}
-}
