+++ src/components/ic_document.jsx
@@ -1,0 +1,18 @@
+import { Component } from 'preact'
+import CSS from './styles.css'
+
+export default class Icon extends Component {
+	render() {
+		let color = this.props.color ? CSS[this.props.color] : CSS.active
+		return (
+			<i className={`${CSS.icon} ${color} ${this.props.className || ''}`}>
+				<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
+					<path
+						fill="#00A0DF"
+						d="M21.414 4.914L17.086 0.586C16.711 0.210901 16.2024 0.000113275 15.672 0L4 0C3.46957 0 2.96086 0.210714 2.58579 0.585786C2.21071 0.960859 2 1.46957 2 2V22C2 22.5304 2.21071 23.0391 2.58579 23.4142C2.96086 23.7893 3.46957 24 4 24H20C20.5304 24 21.0391 23.7893 21.4142 23.4142C21.7893 23.0391 22 22.5304 22 22V6.328C21.9999 5.79761 21.7891 5.28899 21.414 4.914ZM11 19.5H6.5C6.23478 19.5 5.98043 19.3946 5.79289 19.2071C5.60536 19.0196 5.5 18.7652 5.5 18.5C5.5 18.2348 5.60536 17.9804 5.79289 17.7929C5.98043 17.6054 6.23478 17.5 6.5 17.5H11C11.2652 17.5 11.5196 17.6054 11.7071 17.7929C11.8946 17.9804 12 18.2348 12 18.5C12 18.7652 11.8946 19.0196 11.7071 19.2071C11.5196 19.3946 11.2652 19.5 11 19.5ZM15.5 15.5H6.5C6.23478 15.5 5.98043 15.3946 5.79289 15.2071C5.60536 15.0196 5.5 14.7652 5.5 14.5C5.5 14.2348 5.60536 13.9804 5.79289 13.7929C5.98043 13.6054 6.23478 13.5 6.5 13.5H15.5C15.7652 13.5 16.0196 13.6054 16.2071 13.7929C16.3946 13.9804 16.5 14.2348 16.5 14.5C16.5 14.7652 16.3946 15.0196 16.2071 15.2071C16.0196 15.3946 15.7652 15.5 15.5 15.5ZM15.5 11.5H6.5C6.23478 11.5 5.98043 11.3946 5.79289 11.2071C5.60536 11.0196 5.5 10.7652 5.5 10.5C5.5 10.2348 5.60536 9.98043 5.79289 9.79289C5.98043 9.60536 6.23478 9.5 6.5 9.5H15.5C15.7652 9.5 16.0196 9.60536 16.2071 9.79289C16.3946 9.98043 16.5 10.2348 16.5 10.5C16.5 10.7652 16.3946 11.0196 16.2071 11.2071C16.0196 11.3946 15.7652 11.5 15.5 11.5Z"
+					/>
+				</svg>
+			</i>
+		)
+	}
+}
