--- src/components/ic_info.jsx
@@ -1,22 +1,0 @@
-import { Component } from 'preact'
-import CSS from './styles.css'
-
-export default class Icon extends Component {
-	render() {
-		let color = this.props.color ? CSS[this.props.color] : CSS.active
-		return (
-			<i className={`${CSS.icon} ${color} ${this.props.className || ''}`}>
-				<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-					<g fill="none" fillRule="evenodd">
-						<path d="M0 0h24v24H0z" />
-						<path
-							fill="#303334"
-							fillRule="nonzero"
-							d="M12 2c5.52 0 10 4.48 10 10s-4.48 10-10 10S2 17.52 2 12 6.48 2 12 2zm0 18c4.41 0 8-3.59 8-8s-3.59-8-8-8-8 3.59-8 8 3.59 8 8 8zM11 7h2v2h-2V7zm0 4h2v6h-2v-6z"
-						/>
-					</g>
-				</svg>
-			</i>
-		)
-	}
-}
