--- src/components/ic_person.jsx
@@ -1,19 +1,0 @@
-import { Component } from 'preact'
-import CSS from './styles.css'
-
-export default class Icon extends Component {
-	render() {
-		let color = this.props.color ? CSS[this.props.color] : CSS.active
-		return (
-			<i className={`${CSS.icon} ${color} ${this.props.className || ''}`}>
-				<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-					<path
-						fill="#303334"
-						fillRule="nonzero"
-						d="M12 6c1.1 0 2 .9 2 2s-.9 2-2 2-2-.9-2-2 .9-2 2-2zm0 9c2.7 0 5.8 1.29 6 2v1H6v-.99c.2-.72 3.3-2.01 6-2.01zm0-11C9.79 4 8 5.79 8 8s1.79 4 4 4 4-1.79 4-4-1.79-4-4-4zm0 9c-2.67 0-8 1.34-8 4v3h16v-3c0-2.66-5.33-4-8-4z"
-					/>
-				</svg>
-			</i>
-		)
-	}
-}
