+++ src/index.jsx
@@ -16,7 +16,7 @@ export { default as ClockIcon } from './components/ic_clock'
 export { default as CloseIcon } from './components/ic_close'
 export { default as CollapseSmallIcon } from './components/ic_collapse_small'
 export { default as CopyIcon } from './components/ic_copy'
-export { default as DescriptionIcon } from './components/ic_description'
+export { default as DescriptionIcon } from './components/ic_document'
 export { default as DescriptionAddedIcon } from './components/ic_description_added'
 export { default as DownIcon } from './components/ic_down'
 export { default as ExpandSmallIcon } from './components/ic_expand_small'
@@ -29,7 +29,7 @@ export { default as ForwardIcon } from './components/ic_forward'
 export { default as ForwardSmallIcon } from './components/ic_forward_small'
 export { default as GovernmentIcon } from './components/ic_government'
 export { default as HelpIcon } from './components/ic_help'
-export { default as InfoIcon } from './components/ic_info'
+export { default as InfoIcon } from './components/ic_info_bold'
 export { default as KeydeleteIcon } from './components/ic_keydelete'
 export { default as LightIcon } from './components/ic_light'
 export { default as LocationIcon } from './components/ic_location'
@@ -42,7 +42,7 @@ export { default as MopSuccessIcon } from './components/ic_mop_success'
 export { default as MoreIcon } from './components/ic_more'
 export { default as MoveMoneyIcon } from './components/ic_move_money'
 export { default as NotificationsIcon } from './components/ic_notifications'
-export { default as PersonIcon } from './components/ic_person'
+export { default as PersonIcon } from './components/ic_user'
 export { default as PhoneIcon } from './components/ic_phone'
 export { default as QrcardIcon } from './components/ic_qrcard'
 export { default as RecordIcon } from './components/ic_record'
